Places AccessPointProvisioning behind a provisioner interface
As a first step towards introducing a new provisioning method we
need to introduce a new Interface called Provisioner. This allows us
to simply call the Provision() method on structs that implement this
interface, meaning we can expand the supported methods. This also
means we can pull the GidAllocator out of the driver and into the
AccessPointProvisioner which necessitated some test changes.

GidAllocator may well move back into the driver later on as we may
well need to share it across the different provisioners.