Pulled out the GID/UID parsing mechanism into a FileSystemIdentityManager
This way provisioning doesn't need to be concerned with how those the UID/GID
are derived it can just take them on trust. Further it means that dealing
with the GidAllocator is all done in one place and gets rid of lots
of controller tests that were actually testing parsing logic effectively.